EventFlow-AI/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
twilio==8.10.0
assemblyai==0.15.1
requests==2.31.0
gunicorn==21.2.0
python-multipart==0.0.6
